// api/src/api-server/integration/enums.ts
export enum ProxyResponseStatus {
    SUCCESS = 'SUCCESS',
    ERROR = 'ERROR',

}

export enum SupportedPlatforms {
    WOOCOMMERCE = 'WooCommerce',
    PRESTASHOP = 'PrestaShop',
    BASELINKER = 'BaseLinker',
    APILO = 'Apilo',
    EBIUROPROXY = 'EbiuroProxy',
}

export enum IntegrationStatusEnum {
    ACTIVE = 'active',
    INACTIVE = 'inactive',
    PENDING = 'pending',
}

export enum StatusHttp {
    ERROR = 'ERROR',
    SUCCESS = 'SUCCESS',
}